requirements.txt
fastmcp
mcp[cli]
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
httpx
pydantic
typing-extensions
requests